extract ArrayBackedTLongIterator
